chore: improve chain order provider
