+++ packages/widget/src/stores/chains/ChainOrderStore.tsx
@@ -1,10 +1,12 @@
-import { createContext, useContext, useRef } from 'react';
+import { createContext, useContext, useEffect, useRef } from 'react';
+import { useFormContext } from 'react-hook-form';
 import type { StoreApi } from 'zustand';
 import type { UseBoundStoreWithEqualityFn } from 'zustand/traditional';
+import { useChains } from '../../hooks';
+import { FormKey } from '../../providers';
 import type { PersistStoreProviderProps } from '../types';
 import { createChainOrderStore } from './createChainOrderStore';
 import type { ChainOrderState } from './types';
-import { ChainValuesProvider } from './ChainValuesProvider';
 
 export type ChainOrderStore = UseBoundStoreWithEqualityFn<
   StoreApi<ChainOrderState>
@@ -19,13 +21,36 @@ export function ChainOrderStoreProvider({
   ...props
 }: PersistStoreProviderProps) {
   const storeRef = useRef<ChainOrderStore>();
+  const { chains: filteredChains } = useChains();
+  const { setValue, getValues } = useFormContext();
+
   if (!storeRef.current) {
     storeRef.current = createChainOrderStore(props);
   }
 
+  useEffect(() => {
+    if (filteredChains) {
+      const chainOrder = storeRef.current
+        ?.getState()
+        .initializeChains(filteredChains.map((chain) => chain.id));
+      if (chainOrder) {
+        const [fromChainValue, toChainValue] = getValues([
+          FormKey.FromChain,
+          FormKey.ToChain,
+        ]);
+        if (!fromChainValue) {
+          setValue(FormKey.FromChain, chainOrder[0]);
+        }
+        if (!toChainValue) {
+          setValue(FormKey.ToChain, chainOrder[0]);
+        }
+      }
+    }
+  }, [filteredChains, getValues, setValue]);
+
   return (
     <ChainOrderStoreContext.Provider value={storeRef.current}>
-      <ChainValuesProvider>{children}</ChainValuesProvider>
+      {children}
     </ChainOrderStoreContext.Provider>
   );
 }
